Añadir getCards
Servicio para solicitar a la API y modificar getStaticProps en search/cards y search/page para usarlo
